+++ src/modules/tracks/tracks.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common'
-import { AlbumsService } from '../albums/albums.service'
+import { AlbumsService } from '../albums/services/albums.service'
 import { ArtistsService } from '../artists/services/artists.service'
 import { BandsService } from '../bands/services/bands.service'
 import { GenresService } from '../genres/services/genres.service'
